TP Finis

+++ src/Controller/ItemController.php
@@ -32,7 +32,7 @@ class ItemController extends AbstractController
     public function new(Request $request, ItemTypeEntity $itemType = null): Response
     {
         $item = new Item();
-        $form = $this->createForm(ItemType::class, $item);
+        $form = $this->createForm(ItemType::class, $item,['item_type'=>$itemType]);
         $form->handleRequest($request);
 
         if ($form->isSubmitted() && $form->isValid()) {
@@ -54,9 +54,7 @@ class ItemController extends AbstractController
      */
     public function show(Item $item): Response
     {
-        if ($this->isGranted(AppAccess::ITEM_SHOW,$item)===false){
-            return $this->redirectToRoute('item_index');
-        }
+        $this->denyAccessUnlessGranted(AppAccess::ITEM_SHOW,$item);
         return $this->render('item/show.html.twig', ['item' => $item]);
     }
 
@@ -65,6 +63,7 @@ class ItemController extends AbstractController
      */
     public function edit(Request $request, Item $item): Response
     {
+        $this->denyAccessUnlessGranted(AppAccess::ITEM_EDIT,$item);
         $form = $this->createForm(ItemType::class, $item);
         $form->handleRequest($request);
 
@@ -85,6 +84,7 @@ class ItemController extends AbstractController
      */
     public function delete(Request $request, Item $item): Response
     {
+        $this->denyAccessUnlessGranted(AppAccess::ITEM_DELETE,$item);
         if ($this->isCsrfTokenValid('delete'.$item->getId(), $request->request->get('_token'))) {
             $em = $this->getDoctrine()->getManager();
             $em->remove($item);
